UPSTREAM: ALSA: usb-audio: Fix deadlocks at resuming
The recent addition of the USB audio mixer suspend/resume may lead to
deadlocks when the driver tries to call usb_autopm_get_interface()
recursively, since the function tries to sync with the finish of the
other calls.  For avoiding it, introduce a flag indicating the resume
operation and avoids the recursive usb_autopm_get_interface() calls
during the resume.

Reported-and-tested-by: Bryan Quigley <gquigs@gmail.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

BUG=chromium:520917,chromium:530829
TEST=None

Change-Id: I2078d0d0f3daf8a2d928355912765f2cdff3f33a
Signed-off-by: Hsin-Yu Chao <hychao@google.com>
Reviewed-on: https://chromium-review.googlesource.com/300440
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Signed-off-by: MOVZX <movzx@yahoo.com>