Revert "ASoC: compress: Fix compress device direction check"
This breaks sound due to the -EINVAL return; Qualcomm sets both
playback and capture capabilities on the CPU and in the compress
driver. Some people remove just the -EINVAL return; however,
without that block, the patch is useless as playback and capture
are handled the same as without this patch. We don't need it for
our setup so removing it is completely fine.

This reverts commit 5d8add4.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: MOVZX <movzx@yahoo.com>